Minor restructuring toward a more standard project.
With related and incidental documentation updates.
