+++ examples/nuclio/query/query_example_test.go
