+++ pkg/partmgr/partmgr.go
@@ -24,7 +24,7 @@ package partmgr
 
 import (
 	"fmt"
-	"github.com/v3io/v3io-tsdb/config"
+	"github.com/v3io/v3io-tsdb/internal/pkg/config"
 	"github.com/v3io/v3io-tsdb/pkg/aggregate"
 	"sync"
 )
