+++ pkg/tsdbctl/create.go
@@ -22,7 +22,7 @@ package tsdbctl
 
 import (
 	"github.com/spf13/cobra"
-	"github.com/v3io/v3io-tsdb/config"
+	"github.com/v3io/v3io-tsdb/internal/pkg/config"
 	"github.com/v3io/v3io-tsdb/pkg/tsdb"
 )
